IGNITE-18010 Migrate control.sh to IgniteLogger
