+++ modules/control-utility/src/test/java/org/apache/ignite/util/GridCommandHandlerDefragmentationTest.java
@@ -453,7 +453,7 @@ public class GridCommandHandlerDefragmentationTest extends GridCommandHandlerClu
 
     /** */
     private CommandHandler createCommandHandler(ListeningTestLogger testLog) {
-        Logger log = CommandHandler.initLogger(null);
+        Logger log = GridCommandHandlerAbstractTest.initLogger(null);
 
         log.addHandler(new StreamHandler(System.out, new Formatter() {
             /** {@inheritDoc} */
